Changed Event from and to fields to start_at and end_at
The from and to fields were changed to start_at and end_at for postgresql
compatibility.  The forms have been updated for the new fields.  The
locales still call the fields "From" and "To" but are now referenced with
:start_at and :end_at.

The two fields were also change from date to datetime fields.  I have
not included an upgrade migration for either database change.